Include information about where TypeMismatch occurred in exception
Validation would throw a TypeMismatch if a field was not of the expected
type, but there was no way to figure out which field in which struct was
invalid.  This includes this data as fields on the thrown exception.

The message has not been changed to preserve compatibility. A future
revision would include a more informative error message.