Add information about where validation errors occurred to the errors. (#2)
* Move basic (non-strict, required fields only) validation to each write

Previously, the implementation only runs basic, non recursive validation
on the outermost Struct.  This means that nested structs would not have
required fields enforced on serialization.  This moves validation into
writing, which has the effect of making it recursive.  This more closely
matches the original thrift implementation.

* Include information about where TypeMismatch occurred in exception

Validation would throw a TypeMismatch if a field was not of the expected
type, but there was no way to figure out which field in which struct was
invalid.  This includes this data as fields on the thrown exception.

The message has not been changed to preserve compatibility. A future
revision would include a more informative error message.

* Add location information to MissingMandatory errors

Similar to TypeMismatch, MissingMandatory would tell you that a field
was missing, but only hint at the location in the error message.  This
adds in the structured data, while maintaining original error messages
for compatibility.

* Formatting