Remove field question mark methods
These methods like `struct.foo?` answer whether a field has been set.
We don't know whether a field has actually been set, so it just checks
whether the field is nil.  This doesn't seem super useful, and the use
of the question mark is confusing, since in idiomatic ruby, this
predicate form would be a nice way to see whether a boolean field was
true or false.