Annotations on generated class (#22)
* Adding annotations to generated classes

FIELD_ANNOTATIONS and STRUCT_ANNOTATIONS are two to hashes defined for
each generated class. This adds about 1mb to the current build of
jitney-schemas. We want to add this to enable semantic validation by
passing things like regex in the schema definition for string fields

* Use indent_up/down

* Probably should use key instead of name

* Structuring a little different

I think this is a better way of adding the annotations

* Addressed some changes in peer review